Add measurement framerange support
This patch adds the option to specify start and stop frames for
secondary FPS measurements. In addition, options for aborting replay
after the end of the measurement range is reached and for flushing all
work at range boundaries have been added.

The framerange support is very useful when investigating performance for
large traces as creating new trimmed variants for every scene in the
trace can be very time consuming (and quickly becomes unmaintainable for
large sets of traces).
The abort option is very handy on slower development platforms where
replay of relatively small traces can take hours, this allows for
skipping work one might not be interested in for a particular run.

Change-Id: Ic3e81a53974d7955732e65337c2b8669f14ff887